+++ tests/test_cli.py
@@ -289,7 +289,6 @@ def test_combine_lower_version_raises(tmpdir):
         "This error occurs because notebook.py is in the light format in version 55.4, "
         "while jupytext version .* installed at .* can only read the light format in versions .*",
     ):
-
         jupytext([tmp_nbpy, "--to", "ipynb", "--update"])
 
 
@@ -1140,3 +1139,30 @@ def test_jupytext_to_file_emits_a_warning(tmpdir):
 
     with pytest.warns(UserWarning, match="Maybe you want to use the '-o' option"):
         jupytext(["notebook.ipynb", "--to", "script.py"])
+
+
+def test_jupytext_set_formats_file_gives_an_informative_error(tmpdir):
+    """The user may type
+        jupytext --set-formats notebook.md
+    meaning
+        jupytext --syn notebook.md
+    """
+    os.chdir(str(tmpdir))
+
+    cfg_file = tmpdir.join("jupytext.toml")
+    cfg_file.write('default_jupytext_formats = "md,ipynb,py:percent"')
+
+    md_file = tmpdir.join("notebook.md")
+    py_file = tmpdir.join("notebook.py")
+    nb_file = tmpdir.join("notebook.ipynb")
+    md_file.write("Some text")
+
+    with pytest.warns(None) as record:
+        jupytext(["--sync", "notebook.md"])
+
+    assert len(record) == 0
+    assert py_file.exists()
+    assert nb_file.exists()
+
+    with pytest.raises(ValueError, match="jupytext --sync notebook.md"):
+        jupytext(["--set-formats", "notebook.md"])